refactor(types): split events.rs into focused submodules [OPE-369]
Decompose the 754-line monolithic events.rs into:
- kind.rs: AppEventKind enum, Display impl, and query methods
- bus.rs: AppEvent struct and EventBus infrastructure
- tests.rs: complete test suite (87 tests passing)
- mod.rs: public re-exports preserving backwards compatibility
