+++ crates/opengoose-types/src/events/mod.rs
@@ -1,0 +1,8 @@
+mod bus;
+mod kind;
+
+#[cfg(test)]
+mod tests;
+
+pub use bus::{AppEvent, EventBus};
+pub use kind::AppEventKind;
